Reorganize Paste, Recycle, and Delete commands for progress reporting
