+++ Files/Interacts/Interaction.cs
@@ -1,3 +1,4 @@
+using Files.Commands;
 using Files.Dialogs;
 using Files.Enums;
 using Files.Filesystem;
@@ -705,184 +706,7 @@ namespace Files.Interacts
 
         public void DeleteItem_Click(object sender, RoutedEventArgs e)
         {
-            DeleteItem(StorageDeleteOption.Default);
-        }
-
-        public async void DeleteItem(StorageDeleteOption deleteOption)
-        {
-            var deleteFromRecycleBin = App.CurrentInstance.FilesystemViewModel.WorkingDirectory.StartsWith(AppSettings.RecycleBinPath);
-            if (deleteFromRecycleBin)
-            {
-                // Permanently delete if deleting from recycle bin
-                deleteOption = StorageDeleteOption.PermanentDelete;
-            }
-
-            // Get selected items before showing the prompt to prevent deleting items selected after the prompt
-            var CurrentInstance = App.CurrentInstance;
-            List<ListedItem> selectedItems = new List<ListedItem>();
-            foreach (ListedItem selectedItem in CurrentInstance.ContentPage.SelectedItems)
-            {
-                selectedItems.Add(selectedItem);
-            }
-
-            if (AppSettings.ShowConfirmDeleteDialog == true) //check if the setting to show a confirmation dialog is on
-            {
-                var dialog = new ConfirmDeleteDialog(deleteFromRecycleBin, deleteOption);
-                await dialog.ShowAsync();
-
-                if (dialog.Result != MyResult.Delete) //delete selected  item(s) if the result is yes
-                {
-                    return; //return if the result isn't delete
-                }
-                deleteOption = dialog.PermanentlyDelete;
-            }
-            StatusBanner banner = null;
-            try
-            {
-                int itemsDeleted = 0;
-                if (selectedItems.Count > 3)
-                {
-                    if (deleteOption == StorageDeleteOption.PermanentDelete)
-                    {
-                        banner = App.CurrentInstance.StatusBarControl.OngoingTasksControl.PostBanner(null,
-                        CurrentInstance.FilesystemViewModel.WorkingDirectory,
-                        0,
-                        UserControls.StatusBanner.StatusBannerSeverity.Ongoing,
-                        UserControls.StatusBanner.StatusBannerOperation.Delete);
-                    }
-                    else
-                    {
-                        banner = App.CurrentInstance.StatusBarControl.OngoingTasksControl.PostBanner(null,
-                        CurrentInstance.FilesystemViewModel.WorkingDirectory,
-                        0,
-                        UserControls.StatusBanner.StatusBannerSeverity.Ongoing,
-                        UserControls.StatusBanner.StatusBannerOperation.Recycle);
-                    }  
-                }
-                await Task.Run(async () =>
-                {
-                    foreach (ListedItem storItem in selectedItems)
-                    {
-                        uint progressValue = (uint)(itemsDeleted * 100.0 / selectedItems.Count);
-                        await CoreApplication.MainView.CoreWindow.Dispatcher.RunAsync(CoreDispatcherPriority.Normal,
-                            new DispatchedHandler(() =>
-                            {
-                                if (selectedItems.Count > 3) { banner.Report((uint)progressValue); }
-                            }));
-
-                        IStorageItem item;
-                        try
-                        {
-                            if (storItem.PrimaryItemAttribute == StorageItemTypes.File)
-                            {
-                                item = await ItemViewModel.GetFileFromPathAsync(storItem.ItemPath);
-                            }
-                            else
-                            {
-                                item = await ItemViewModel.GetFolderFromPathAsync(storItem.ItemPath);
-                            }
-
-                            await item.DeleteAsync(deleteOption);
-                        }
-                        catch (UnauthorizedAccessException)
-                        {
-                            if (deleteOption == StorageDeleteOption.Default)
-                            {
-                                // Try again with fulltrust process
-                                if (App.Connection != null)
-                                {
-                                    var result = await App.Connection.SendMessageAsync(new ValueSet() {
-                                    { "Arguments", "FileOperation" },
-                                    { "fileop", "MoveToBin" },
-                                    { "filepath", storItem.ItemPath } });
-                                }
-                            }
-                            else
-                            {
-                                // Try again with DeleteFileFromApp
-                                if (!NativeDirectoryChangesHelper.DeleteFileFromApp(storItem.ItemPath))
-                                {
-                                    Debug.WriteLine(System.Runtime.InteropServices.Marshal.GetLastWin32Error());
-                                }
-                            }
-                        }
-                        catch (FileLoadException)
-                        {
-                            // try again
-                            if (storItem.PrimaryItemAttribute == StorageItemTypes.File)
-                            {
-                                item = await ItemViewModel.GetFileFromPathAsync(storItem.ItemPath);
-                            }
-                            else
-                            {
-                                item = await ItemViewModel.GetFolderFromPathAsync(storItem.ItemPath);
-                            }
-
-                            await item.DeleteAsync(deleteOption);
-                        }
-
-                        if (deleteFromRecycleBin)
-                        {
-                            // Recycle bin also stores a file starting with $I for each item
-                            var iFilePath = Path.Combine(Path.GetDirectoryName(storItem.ItemPath), Path.GetFileName(storItem.ItemPath).Replace("$R", "$I"));
-                            await (await ItemViewModel.GetFileFromPathAsync(iFilePath)).DeleteAsync(StorageDeleteOption.PermanentDelete);
-                        }
-
-                        CurrentInstance.FilesystemViewModel.RemoveFileOrFolder(storItem);
-                        itemsDeleted++;
-                    }
-                });
-                if (deleteOption == StorageDeleteOption.PermanentDelete)
-                {
-                    App.CurrentInstance.StatusBarControl.OngoingTasksControl.PostBanner(
-                    "Deletion Complete",
-                    "The operation has completed.",
-                    0,
-                    StatusBanner.StatusBannerSeverity.Success,
-                    StatusBanner.StatusBannerOperation.Delete);
-                }
-                else
-                {
-                    App.CurrentInstance.StatusBarControl.OngoingTasksControl.PostBanner(
-                    "Recycle Complete",
-                    "The operation has completed.",
-                    0,
-                    StatusBanner.StatusBannerSeverity.Success,
-                    StatusBanner.StatusBannerOperation.Recycle);
-                }
-                
-                App.CurrentInstance.NavigationToolbar.CanGoForward = false;
-            }
-            catch (UnauthorizedAccessException)
-            {
-                if (StatusCenter.StatusBannersSource.Contains(banner)) StatusCenter.StatusBannersSource.Remove(banner);
-                App.CurrentInstance.StatusBarControl.OngoingTasksControl.PostBanner(
-                    ResourceController.GetTranslation("AccessDeniedDeleteDialog/Title"),
-                    ResourceController.GetTranslation("AccessDeniedDeleteDialog/Text"),
-                    0,
-                    StatusBanner.StatusBannerSeverity.Error,
-                    StatusBanner.StatusBannerOperation.Delete);
-            }
-            catch (FileNotFoundException)
-            {
-                if (StatusCenter.StatusBannersSource.Contains(banner)) StatusCenter.StatusBannersSource.Remove(banner);
-                App.CurrentInstance.StatusBarControl.OngoingTasksControl.PostBanner(
-                    ResourceController.GetTranslation("FileNotFoundDialog/Title"),
-                    ResourceController.GetTranslation("FileNotFoundDialog/Text"),
-                    0,
-                    StatusBanner.StatusBannerSeverity.Error,
-                    StatusBanner.StatusBannerOperation.Delete);
-            }
-            catch (IOException)
-            {
-                if (StatusCenter.StatusBannersSource.Contains(banner))  StatusCenter.StatusBannersSource.Remove(banner);
-                App.CurrentInstance.StatusBarControl.OngoingTasksControl.PostActionBanner(
-                    ResourceController.GetTranslation("FileInUseDeleteDialog/Title"),
-                    ResourceController.GetTranslation("FileInUseDeleteDialog/Text"),
-                    ResourceController.GetTranslation("FileInUseDeleteDialog/PrimaryButtonText"),
-                    ResourceController.GetTranslation("FileInUseDeleteDialog/SecondaryButtonText"), () => { DeleteItem(deleteOption); });
-            }
-            App.CurrentInstance.StatusBarControl.OngoingTasksControl.RemoveBanner(banner);
+            ItemOperations.DeleteItemWithStatus(StorageDeleteOption.Default);
         }
 
         public void RenameItem_Click(object sender, RoutedEventArgs e)
@@ -1151,8 +975,6 @@ namespace Files.Interacts
         }
 
         public string CopySourcePath;
-        public IReadOnlyList<IStorageItem> itemsToPaste;
-        public int itemsPasted;
 
         public async void CopyItem_ClickAsync(object sender, RoutedEventArgs e)
         {
@@ -1213,12 +1035,6 @@ namespace Files.Interacts
             }
         }
 
-        private enum ImpossibleActionResponseTypes
-        {
-            Skip,
-            Abort
-        }
-
         public async void EmptyRecycleBin_ClickAsync(object sender, RoutedEventArgs e)
         {
             var ConfirmEmptyBinDialog = new ContentDialog()
@@ -1244,277 +1060,12 @@ namespace Files.Interacts
             }
         }
 
-        public async void PasteItem_ClickAsync(object sender, RoutedEventArgs e)
+        public void PasteItem_ClickAsync(object sender, RoutedEventArgs e)
         {
             DataPackageView packageView = Clipboard.GetContent();
             string destinationPath = CurrentInstance.FilesystemViewModel.WorkingDirectory;
 
-            await PasteItems(packageView, destinationPath, packageView.RequestedOperation);
-        }
-
-        public async Task PasteItems(DataPackageView packageView, string destinationPath, DataPackageOperation acceptedOperation)
-        {
-            if (!packageView.Contains(StandardDataFormats.StorageItems))
-            {
-                // Happens if you copy some text and then you Ctrl+V in FilesUWP
-                // Should this be done in ModernShellPage?
-                return;
-            }
-            if (CurrentInstance.FilesystemViewModel.WorkingDirectory.StartsWith(AppSettings.RecycleBinPath))
-            {
-                // Do not paste files and folders inside the recycle bin
-                await DialogDisplayHelper.ShowDialog(ResourceController.GetTranslation("ErrorDialogThisActionCannotBeDone"), ResourceController.GetTranslation("ErrorDialogUnsupportedOperation"));
-                return;
-            }
-
-            itemsToPaste = await packageView.GetStorageItemsAsync();
-            HashSet<IStorageItem> pastedSourceItems = new HashSet<IStorageItem>();
-            HashSet<IStorageItem> pastedItems = new HashSet<IStorageItem>();
-            itemsPasted = 0;
-            StatusBanner banner = null;
-            if (itemsToPaste.Count > 3)
-            {
-                banner = App.CurrentInstance.StatusBarControl.OngoingTasksControl.PostBanner(null,
-                        CurrentInstance.FilesystemViewModel.WorkingDirectory,
-                        0,
-                        StatusBanner.StatusBannerSeverity.Ongoing,
-                        StatusBanner.StatusBannerOperation.Paste);
-            }
-
-            await Task.Run(async () =>
-            {
-                foreach (IStorageItem item in itemsToPaste)
-                {
-                    if (item.IsOfType(StorageItemTypes.Folder))
-                    {
-                        if (!string.IsNullOrEmpty(item.Path) && destinationPath.IsSubPathOf(item.Path))
-                        {
-                            ImpossibleActionResponseTypes responseType = ImpossibleActionResponseTypes.Abort;
-
-                            /// Currently following implementation throws exception until it is resolved keep it disabled
-                            /*Binding themeBind = new Binding();
-                            themeBind.Source = ThemeHelper.RootTheme;
-
-                            ContentDialog dialog = new ContentDialog()
-                            {
-                                Title = ResourceController.GetTranslation("ErrorDialogThisActionCannotBeDone"),
-                                Content = ResourceController.GetTranslation("ErrorDialogTheDestinationFolder") + " (" + destinationPath.Split(Path.DirectorySeparatorChar, StringSplitOptions.RemoveEmptyEntries).Last() + ") " + ResourceController.GetTranslation("ErrorDialogIsASubfolder") + " (" + item.Name + ")",
-                                PrimaryButtonText = ResourceController.GetTranslation("ErrorDialogSkip"),
-                                CloseButtonText = ResourceController.GetTranslation("ErrorDialogCancel"),
-                                PrimaryButtonCommand = new RelayCommand(() => { responseType = ImpossibleActionResponseTypes.Skip; }),
-                                CloseButtonCommand = new RelayCommand(() => { responseType = ImpossibleActionResponseTypes.Abort; }),
-                            };
-                            BindingOperations.SetBinding(dialog, FrameworkElement.RequestedThemeProperty, themeBind);
-
-                            await dialog.ShowAsync();*/
-                            if (responseType == ImpossibleActionResponseTypes.Skip)
-                            {
-                                continue;
-                            }
-                            else if (responseType == ImpossibleActionResponseTypes.Abort)
-                            {
-                                return;
-                            }
-                        }
-                        else
-                        {
-                            try
-                            {
-                                ClonedDirectoryOutput pastedOutput = await CloneDirectoryAsync(
-                                    (StorageFolder)item,
-                                    await ItemViewModel.GetFolderFromPathAsync(destinationPath),
-                                    item.Name, false, banner);
-                                pastedSourceItems.Add(item);
-                                pastedItems.Add(pastedOutput.FolderOutput);
-                                await CoreApplication.MainView.CoreWindow.Dispatcher.RunAsync(CoreDispatcherPriority.Normal,
-                                    new DispatchedHandler(() =>
-                                    {
-                                        banner = pastedOutput.StatusBannerOutput;
-                                    }));
-                            }
-                            catch (FileNotFoundException)
-                            {
-                                // Folder was moved/deleted in the meantime
-                                continue;
-                            }
-                        }
-                    }
-                    else if (item.IsOfType(StorageItemTypes.File))
-                    {
-                        uint progressValue = (uint)(itemsPasted * 100.0 / itemsToPaste.Count);
-                        if (itemsToPaste.Count > 3)
-                        {
-                            await CoreApplication.MainView.CoreWindow.Dispatcher.RunAsync(CoreDispatcherPriority.Normal,
-                                new DispatchedHandler(() =>
-                                {
-                                    banner.Report((uint)progressValue);
-                                }));
-                        }
-                        try
-                        {
-                            StorageFile clipboardFile = (StorageFile)item;
-                            StorageFile pastedFile = await clipboardFile.CopyAsync(
-                                await ItemViewModel.GetFolderFromPathAsync(destinationPath),
-                                item.Name,
-                                NameCollisionOption.GenerateUniqueName);
-                            pastedSourceItems.Add(item);
-                            pastedItems.Add(pastedFile);
-                        }
-                        catch (UnauthorizedAccessException)
-                        {
-                            // Try again with CopyFileFromApp
-                            if (NativeDirectoryChangesHelper.CopyFileFromApp(item.Path, Path.Combine(destinationPath, item.Name), true))
-                            {
-                                pastedSourceItems.Add(item);
-                            }
-                            else
-                            {
-                                Debug.WriteLine(System.Runtime.InteropServices.Marshal.GetLastWin32Error());
-                            }
-                        }
-                        catch (FileNotFoundException)
-                        {
-                            // File was moved/deleted in the meantime
-                            continue;
-                        }
-                    }
-                    itemsPasted++;
-                }
-
-                if (acceptedOperation == DataPackageOperation.Move)
-                {
-                    foreach (IStorageItem item in pastedSourceItems)
-                    {
-                        try
-                        {
-                            if (string.IsNullOrEmpty(item.Path))
-                            {
-                                // Can't move (only copy) files from MTP devices because:
-                                // StorageItems returned in DataPackageView are read-only
-                                // The item.Path property will be empty and there's no way of retrieving a new StorageItem with R/W access
-                                continue;
-                            }
-                            if (item.IsOfType(StorageItemTypes.File))
-                            {
-                                // If we reached this we are not in an MTP device, using StorageFile.* is ok here
-                                StorageFile file = await StorageFile.GetFileFromPathAsync(item.Path);
-                                await file.DeleteAsync(StorageDeleteOption.PermanentDelete);
-                            }
-                            else if (item.IsOfType(StorageItemTypes.Folder))
-                            {
-                                // If we reached this we are not in an MTP device, using StorageFolder.* is ok here
-                                StorageFolder folder = await StorageFolder.GetFolderFromPathAsync(item.Path);
-                                await folder.DeleteAsync(StorageDeleteOption.PermanentDelete);
-                            }
-                        }
-                        catch (UnauthorizedAccessException)
-                        {
-                            // Try again with DeleteFileFromApp
-                            if (!NativeDirectoryChangesHelper.DeleteFileFromApp(item.Path))
-                            {
-                                Debug.WriteLine(System.Runtime.InteropServices.Marshal.GetLastWin32Error());
-                            }
-                        }
-                        catch (FileNotFoundException)
-                        {
-                            // File or Folder was moved/deleted in the meantime
-                            continue;
-                        }
-                        ListedItem listedItem = CurrentInstance.FilesystemViewModel.FilesAndFolders.FirstOrDefault(listedItem => listedItem.ItemPath.Equals(item.Path, StringComparison.OrdinalIgnoreCase));
-                    }
-                }
-            });
-
-            if (destinationPath == CurrentInstance.FilesystemViewModel.WorkingDirectory)
-            {
-                List<string> pastedItemPaths = pastedItems.Select(item => item.Path).ToList();
-                List<ListedItem> copiedItems = CurrentInstance.FilesystemViewModel.FilesAndFolders.Where(listedItem => pastedItemPaths.Contains(listedItem.ItemPath)).ToList();
-                if (copiedItems.Any())
-                {
-                    CurrentInstance.ContentPage.SetSelectedItemsOnUi(copiedItems);
-                    CurrentInstance.ContentPage.FocusSelectedItems();
-                }
-            }
-            packageView.ReportOperationCompleted(acceptedOperation);
-            App.CurrentInstance.StatusBarControl.OngoingTasksControl.RemoveBanner(banner);
-        }
-
-        public class ClonedDirectoryOutput
-        {
-            public StorageFolder FolderOutput { get; set; }
-            public StatusBanner StatusBannerOutput { get; set; } = null;
-        }
-
-        public async Task<ClonedDirectoryOutput> CloneDirectoryAsync(StorageFolder SourceFolder, StorageFolder DestinationFolder, string sourceRootName, bool suppressProgressFlyout, StatusBanner banner = null)
-        {
-            var createdRoot = await DestinationFolder.CreateFolderAsync(sourceRootName, CreationCollisionOption.GenerateUniqueName);
-            DestinationFolder = createdRoot;
-
-            foreach (StorageFile fileInSourceDir in await SourceFolder.GetFilesAsync())
-            {
-                if (itemsToPaste != null)
-                {
-                    if (itemsToPaste.Count > 3 && !suppressProgressFlyout)
-                    {
-                        uint progressValue = (uint)(itemsPasted * 100.0 / (itemsToPaste.Count + (await SourceFolder.GetFilesAsync()).Count));
-                        await CoreApplication.MainView.CoreWindow.Dispatcher.RunAsync(CoreDispatcherPriority.Normal,
-                            new DispatchedHandler(() =>
-                            {
-                                if (banner == null)
-                                {
-                                    throw new ArgumentNullException();
-                                }
-                                else
-                                {
-                                    banner.Report(progressValue);
-                                }
-                            }));
-                    }
-                }
-
-                await fileInSourceDir.CopyAsync(DestinationFolder, fileInSourceDir.Name, NameCollisionOption.GenerateUniqueName);
-            }
-            foreach (StorageFolder folderinSourceDir in await SourceFolder.GetFoldersAsync())
-            {
-                if (itemsToPaste != null)
-                {
-                    if (itemsToPaste.Count > 3 && !suppressProgressFlyout)
-                    {
-                        uint progressValue = (uint)(itemsPasted * 100.0 / (itemsToPaste.Count + (await SourceFolder.GetFoldersAsync()).Count));
-                        await CoreApplication.MainView.CoreWindow.Dispatcher.RunAsync(CoreDispatcherPriority.Normal,
-                            new DispatchedHandler(() =>
-                            {
-                                if (banner == null)
-                                {
-                                    throw new ArgumentNullException();
-                                }
-                                else
-                                {
-                                    banner.Report(progressValue);
-                                }
-                            }));
-                    }
-                }
-
-                var output = await CloneDirectoryAsync(folderinSourceDir, DestinationFolder, folderinSourceDir.Name, false, banner);
-                banner = output.StatusBannerOutput;
-            }
-            if (!suppressProgressFlyout)
-            {
-                return new ClonedDirectoryOutput()
-                {
-                    FolderOutput = createdRoot,
-                    StatusBannerOutput = banner
-                };
-            }
-            else
-            {
-                return new ClonedDirectoryOutput()
-                {
-                    FolderOutput = createdRoot,
-                    StatusBannerOutput = null
-                };
-            }
+            ItemOperations.PasteItemWithStatus(packageView, destinationPath, packageView.RequestedOperation);
         }
 
         public void NewFolder_Click(object sender, RoutedEventArgs e)
