+++ Files/UserControls/MultitaskingControl/VerticalTabViewControl.xaml.cs
@@ -1,4 +1,5 @@
-﻿using Files.Common;
+﻿using Files.Commands;
+using Files.Common;
 using Files.Interacts;
 using Files.UserControls.MultiTaskingControl;
 using Files.Views;
@@ -214,7 +215,7 @@ namespace Files.UserControls
             }
         }
 
-        private async void TabViewItem_Drop(object sender, DragEventArgs e)
+        private void TabViewItem_Drop(object sender, DragEventArgs e)
         {
             if (e.DataView.AvailableFormats.Contains(StandardDataFormats.StorageItems))
             {
@@ -227,18 +228,7 @@ namespace Files.UserControls
                     .FilesystemViewModel
                     .WorkingDirectory;
 
-                foreach (IStorageItem item in await e.DataView.GetStorageItemsAsync())
-                {
-                    if (item.IsOfType(StorageItemTypes.Folder))
-                    {
-                        await App.CurrentInstance.InteractionOperations.CloneDirectoryAsync((StorageFolder)item, await StorageFolder.GetFolderFromPathAsync(tabViewItemWorkingDir), item.Name, true);
-                        await (await StorageFolder.GetFolderFromPathAsync(item.Path)).DeleteAsync();
-                    }
-                    else
-                    {
-                        await ((StorageFile)item).MoveAsync(await StorageFolder.GetFolderFromPathAsync(tabViewItemWorkingDir), item.Name, NameCollisionOption.GenerateUniqueName);
-                    }
-                }
+                ItemOperations.PasteItemWithStatus(e.DataView, tabViewItemWorkingDir, DataPackageOperation.Move);
             }
             else
             {
